+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
  * Бандл для интеграции с Symfony фреймфорком переехал в отдельный 
    [репозиторий](https://github.com/biplane/BiplaneYandexDirectBundle).
+   
+ * `Biplane\YandexDirect\Api\V4SoapClient` переименован в `Biplane\YandexDirect\Api\SoapClientV4`,
+   `Biplane\YandexDirect\Api\V5SoapClient` переименован в `Biplane\YandexDirect\Api\SoapClientV5`.
 
 ## 3.4.2 [commit logs](https://github.com/biplane/BiplaneYandexDirectBundle/compare/3.4.1...3.4.2)